requirements.txt
filemagic
PyPDF2
pycrypto
requests
python-documentcloud
google-api-python-client